Feature 6
* creacion , eliminacion y obtencion de usuarios lista

+++ src/models/mySql/notesModel.ts
@@ -1,6 +1,7 @@
 import mysql, { Connection } from 'mysql2/promise'
 
 import { UUID, UserModel } from '../../types'
+import { RowDataPacket } from 'mysql2'
 
 const DEFAULT_CONFIG = {
   host: 'localhost',
@@ -29,51 +30,62 @@ export class notesModel {
     }
   }
 
-  static async createUser ({ data }: { data: UserModel }): Promise<any> {
+  static async createUser ({ data }: { data: UserModel }): Promise<RowDataPacket[][]> {
     const { name, email, telefono } = data
     const connectiondb = await connect()
-
-    const [uuidResult]: any = await connectiondb?.query('select uuid() uuid')
-    const [{ uuid }] = uuidResult
-    try {
-      await connectiondb?.query(
-        'insert into task_glide.usuarios values (UUID_TO_BIN(?),?,?,?)',
-        [uuid, name, email, telefono]
+    if (connectiondb != null) {
+      const [uuidResult] = await connectiondb.query<RowDataPacket[]>(
+        'select uuid() uuid'
       )
-    } catch (e) {
-      throw new Error('Error al crear el usuario')
-    }
-    const [result]: any = await connectiondb?.query(
-      'select  nombre , email ,telefono from task_glide.usuarios;'
-    )
-    return result
-  }
-
-  static async getAllUser (): Promise<void> {
-    const connectiondb = await connect()
-    try {
-      const [result] = await connectiondb?.query(
-        'select  BIN_TO_UUID(id_usuario) id_usuario,nombre , email ,telefono from task_glide.usuarios;'
+      const [{ uuid }] = uuidResult
+      try {
+        const query: string =
+          'insert into task_glide.usuarios values (UUID_TO_BIN(?),?,?,?)'
+        await connectiondb?.query(query, [uuid, name, email, telefono])
+      } catch (e) {
+        throw new Error('Error al crear el usuario')
+      }
+      const [result] = await connectiondb.query<RowDataPacket[][]>(
+        'select  nombre , email ,telefono from task_glide.usuarios;'
       )
       return result
-    } catch (e) {
-      throw new Error('Error al consultar el usuario')
+    } else {
+      throw new Error('Error al conectar con la base de datos')
     }
   }
 
-  static async deleteUser ({ id }: { id: UUID }): Promise<void> {
+  static async getAllUser (): Promise<RowDataPacket[][]> {
     const connectiondb = await connect()
-    try {
-      await connectiondb?.query(
-        'delete from task_glide.usuarios where usuarios.id_usuario = UUID_TO_BIN(?)',
-        id
-      )
-    } catch (e) {
-      throw new Error('Error al eliminar el usuario')
+    if (connectiondb != null) {
+      try {
+        const query: string =
+          'select  BIN_TO_UUID(id_usuario) id_usuario,nombre , email ,telefono from task_glide.usuarios;'
+        const [result] = await connectiondb.query<RowDataPacket[][]>(query)
+        return result
+      } catch (e) {
+        throw new Error('Error al consultar el usuario')
+      }
+    } else {
+      throw new Error('Error al conectar con la base de datos')
     }
-    const [result]: any = await connectiondb?.query(
-      'select  nombre , email ,telefono from task_glide.usuarios;'
-    )
-    return result
+  }
+
+  static async deleteUser ({ id }: { id: UUID }): Promise<RowDataPacket[][]> {
+    const connectiondb = await connect()
+    if (connectiondb != null) {
+      try {
+        const query: string =
+          'delete from task_glide.usuarios where usuarios.id_usuario = UUID_TO_BIN(?);'
+        await connectiondb.query(query, id)
+      } catch (e) {
+        throw new Error('Error al eliminar el usuario')
+      }
+      const [result] = await connectiondb.query<RowDataPacket[][]>(
+        'select  nombre , email ,telefono from task_glide.usuarios;'
+      )
+      return result
+    } else {
+      throw new Error('error al conectar con la base de datos')
+    }
   }
 }
